third_party/WebKit/LayoutTests/inspector-protocol/heap-profiler/heap-samples-in-snapshot.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

function junkGenerator()
{
    var junkArray = new Array(1000);
    for (var i = 0; i < junkArray.length; ++i)
        junkArray[i] = "42 " + i;
    window.junkArray = junkArray;
}

function setupIntervalAndRunTest()
{
    setInterval(junkGenerator, 0);
    runTest();
}

function test()
{
    InspectorTest.importScript("../../../../inspector-protocol/heap-profiler/resources/heap-snapshot-common.js");

    function trackingStarted()
    {
        InspectorTest.log("Tracking started");
    }

    function arraysEqual(a, b)
    {
        var sa = a.join(", ");
        var sb = b.join(", ");
        if (sa === sb)
            return true;
        InspectorTest.log("FAIL:\na = [" + sa+ "]\nb = [" + sb + "]");
        return false;
    }

    function trackingStopped(snapshot)
    {
        var samples = snapshot.getSamples();
        InspectorTest.log("Last assigned id arrays match: " + arraysEqual(lastAssignedIds, samples.lastAssignedIds));
        var sizesMatch = (sizes.length <= samples.sizes.length);
        InspectorTest.log("Size arrays length is correct: " + sizesMatch);
        if (!sizesMatch) {
            // Print mismatch:
            arraysEqual(sizes, samples.sizes);
        }
        var sizesNonGrowing = true;
        for (var i = 0; i < samples.sizes.length; i++) {
            if ((sizes[i] === undefined && samples.sizes[i] !== 0) || (sizes[i] < samples.sizes[i])) {
                sizesNonGrowing = false;
                InspectorTest.log("FAIL: total size of live objects from interval cannot increase.");
                // Print mismatch:
                arraysEqual(sizes, samples.sizes);
            }
        }
        InspectorTest.log("Sizes non growing: " + sizesNonGrowing);
        InspectorTest.completeTest();
    }

    var sizes = [];
    var lastAssignedIds = [];
    InspectorTest.eventHandler["HeapProfiler.lastSeenObjectId"] = function(messageObject)
    {
        lastAssignedIds.push(messageObject["params"]["lastSeenObjectId"]);
        if (lastAssignedIds.length === 2) {
            // Wait for two updates and then stop tracing.
            InspectorTest.stopRecordingHeapTimeline(trackingStopped);
        }
    }

    InspectorTest.eventHandler["HeapProfiler.heapStatsUpdate"] = function(messageObject)
    {
        var samples = messageObject["params"]["statsUpdate"];
        for (var i = 0; i < samples.length; i += 3) {
            index = samples[i];
            sizes[index] = samples[i+2];
        }
    }

    InspectorTest.sendCommand("HeapProfiler.startTrackingHeapObjects", {}, trackingStarted);
    //# sourceURL=heap-objects-tracking.html
}
</script>
</head>
<body onload="setupIntervalAndRunTest()">
<p>Test that heap tracking actually reports data fragments.</p>
</body>
</html>
