third_party/blink/web_tests/platform/generic/inspector-protocol/heap-profiler/heap-samples-in-snapshot-expected.txt

Test that heap tracking actually reports data fragments.
Tracking started
Took heap snapshot
Parsed snapshot
Last assigned id arrays match: true
Size arrays length is correct: true
Sizes non growing: true
